+++ lib/cpuinfo.dart
@@ -5,9 +5,8 @@ import 'minMaxFreq.dart';
 class CpuInfo {
   int numberOfCores;
   String abi;
-  Map<String, MinMaxFrequency> minMaxFrequencies =
-      Map<String, MinMaxFrequency>();
-  Map<String, int> currentFriquencies;
+  Map<int, MinMaxFrequency> minMaxFrequencies = Map<int, MinMaxFrequency>();
+  Map<int, int> currentFriquencies = Map<int, int>();
 
   CpuInfo(
       {this.numberOfCores,
@@ -16,25 +15,28 @@ class CpuInfo {
       this.currentFriquencies});
 
   // Deserialize the data retrieved from the device through platform specific code
-  CpuInfo.fromJson(Map<String, dynamic> json) {
+  CpuInfo.fromJson(Map<dynamic, dynamic> json) {
     this.numberOfCores = json['numberOfCores'];
     this.abi = json['abi'];
-    this.currentFriquencies = Map.from(json['currentFrequencies']);
+    Map.from(json['currentFrequencies']).forEach((key, value) {
+      this.currentFriquencies[int.parse(key)] = value;
+    });
 
     Map.from(json['minMaxFrequencies']).forEach((key, value) {
       var map = Map.from(value);
-      this.minMaxFrequencies[key] =
+      this.minMaxFrequencies[int.parse(key)] =
           MinMaxFrequency(map['first'], map['second']);
     });
   }
 
-  // Serialize the data to json
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['abi'] = this.abi;
-    data['numberOfCores'] = this.numberOfCores;
-    data['currentFrequencies'] = this.currentFriquencies;
-    data['minMaxFrequencies'] = this.minMaxFrequencies;
-    return data;
+  Map<String, dynamic> toJson() => {
+        "abi": abi,
+        "numberOfCores": numberOfCores,
+        "currentFrequencies": currentFriquencies.map(convert),
+        "minMaxFrequencies": minMaxFrequencies.map(convert)
+      };
+
+  MapEntry convert<T>(int key, T value) {
+    return MapEntry(key.toString(), value);
   }
 }